+++ R/Rdimtools-package.R
@@ -38,6 +38,7 @@
 #' \code{\link{do.extlpp}} \tab U \tab Extended Locality Preserving Projection \cr
 #' \code{\link{do.fa}} \tab U \tab (Exploratory) Factor Analysis  \cr
 #' \code{\link{do.fscore}} \tab S \tab Fisher Score \cr
+#' \code{\link{do.fssem}} \tab U \tab Feature Subset Selection using Expectation-Maximization \cr
 #' \code{\link{do.ica}} \tab U \tab Independent Component Analysis \cr
 #' \code{\link{do.isoproj}} \tab U \tab Isometric Projection \cr
 #' \code{\link{do.kmvp}} \tab S \tab Kernel-Weighted Maximum Variance Projection \cr
@@ -112,6 +113,7 @@
 #' Also, we have \emph{nonlinear} methods implemented
 #' \tabular{lcl}{
 #' FUNCTION \tab TYPE \tab ALGORITHM \cr
+#' \code{\link{do.cge}} \tab SS \tab Constrained Graph Embedding \cr
 #' \code{\link{do.cisomap}} \tab U \tab Conformal Isometric Feature Mapping \cr
 #' \code{\link{do.crca}} \tab U \tab Curvilinear Component Analysis \cr
 #' \code{\link{do.crda}} \tab U \tab Curvilinear Distance Analysis \cr
@@ -205,6 +207,7 @@
 #' @import Rdpack
 #' @import CVXR
 #' @import RcppDE
+#' @importFrom ClusterR Optimal_Clusters_GMM GMM predict_GMM
 #' @importFrom geigen geigen
 #' @importFrom utils packageVersion combn
 #' @importFrom RSpectra eigs svds