+++ R/linear_PCA.R
@@ -46,7 +46,7 @@
 #' plot(out3$Y[,1], out3$Y[,2], main="PCA::'whiten'")
 #' }
 #'
-#' @author Changhee Suh
+#' @author Kisung You
 #' @references
 #' \insertRef{pearson_liii._1901}{Rdimtools}
 #'