+++ NEWS.md
@@ -1,9 +1,12 @@
 # News for Package Rdimtools
 ### Changes in version 0.4.3
+  * NEW functions are added
+    - `do.fssem` : Feature Subset Selection using Expectation-Maximization.
+    - `do.cge`   : Constrained Graph Embedding.
   * LDA solves trace ratio problem directly.
   * Supports
-    - 125 dimension reduction methods
-    - 13  intrinsic dimension estimation techniques 
+    - 125 dimension reduction methods before this version.
+    - 13  intrinsic dimension estimation techniques.
   * Initialize the following documentation:
     - NEWS for brief record of updates.
     - README for introduction.